src/main/java/springbook/user/service/UserService.java
package springbook.user.service;

import org.springframework.mail.MailSender;
import org.springframework.mail.SimpleMailMessage;
import org.springframework.transaction.PlatformTransactionManager;
import org.springframework.transaction.TransactionStatus;
import org.springframework.transaction.support.DefaultTransactionDefinition;
import springbook.user.dao.UserDao;
import springbook.user.domain.Level;
import springbook.user.domain.User;

import java.sql.SQLException;
import java.util.List;

/**
 * Created by adaeng on 11/03/2019.


 */
public interface UserService {

    public void upgradeLevels() throws SQLException;


    public void add(User user);


}
